STM: always stock in vio files the first node (state) of a proof
It used not to be the case when the proof contains Sideff, since
the code was picking the last known state and not necessarily the
first one.  Because of side effects the last known state could be
the one corresponding to the side effect (that was executed to, say,
change the parser).  This is also related to bug #4530