[search] Don't build intermediate lists in search.
This patch converts the `search_*` functions to use an iter-style API.

Consequently, the Search Vernac will produce a message for each search
result, greatly improving roundtrip time as IDEs can effectively display
the results in a streaming way.

It also allows different printers to be used.

I didn't observe a performance difference (as things seem to be
dominated by printing and `Declaremods`).

As a minor tweak, we make search with "Output Name Only" more efficient.

Motivation:
-----------

Currently, the main search API `Search.generic_search` is an effectful,
iteration-based function:

```ocaml
val generic_search : int option -> display_function -> unit
```

This type is imposed by `Declaremods`, which exposes an effectful,
iteration-based API to traverse Coq library objects.

The `Search.search_*` functions try to offer a more functional API by
returning a list of pretty printing commands. They need to build an
internal intermediate list for that purpose.

However, this is a waste of time, as the destination of these lists is
to be flushed out by the printer right away.