CoqIDE is more resilient to initialization errors.
We force the STM to finish after the initialization request so as to raise
exceptions that may have been generated by the initialization process. Likewise,
we simply die when the initialization request fails in CoqIDE instead of just
printing an error message.

This is the fix for the underlying issue of bug #4591, but it does not solve
the bug yet.