Revert "refine: do check all unif problems are solved (Close: #4415, #4532)"
This fix is too restrictive.  Still, opening a goal for an evar with a
pending conv_pb is unsafe since the user may prove (instantiate it) in
a way not compatible with the conv_pb.

Assigning an evar, in its lowest level API, should enforce that all
related conv_pbs are satisfied by the instance.
This also poses a UI problem, since there is not way to see these
conv_pbs.  One could print a goal and say: look, the proof term you give
must validate this equation...

Given that the good fix is not obvious, we revert!

This reverts commit a0e792236c9666df1069753f8f807c12f713dcfb.