Removing the ad-hoc tactic_expr type.
This type was actually only used by the debug printer of tactics, and only
for atomic tactics. Furthermore, that type was asymmetric, as the underlying
tacexpr type was set to be glob_tactic, when the semantics would have required
a Val.t type.

Furthermore, this type is absent from every contrib I have seen, which hints
again in favour of its lack of meaning.