A fix to #3709: ensuring extra parentheses when a tactic entry has a subentry at a higher tactic level than the entry itself.
This is applicable to the parsing of expressions with infix or postfix
operators such as ; or ||.

Could be improved, e.g. so that no parenthesis are put when the
expression is the rightmost one, as in:

  now (tac1;tac2)

where parentheses are not needed but still printed with this patch,
while the patch adds needed parentheses in

  (now tac1);tac2

This would hardly scale to more complex grammars. E.g., if a suffix
expression can extend a leading expression as part of different
grammar entries, as in

  let x := simpl in y ...

I don't see in general how to anticipate the need for parentheses
without restarting the parser to check the reversibility of the
printing.