coqc: support -o option to specify output file name
The -o option lets one put .vo or .vio files in a directory of choice,
i.e. decouple the location of the sources and the compiled files.
This ease the integration of Coq in already existing IDEs that handle
the build process automatically (eg Eclipse) and also enables one to
compile/run at the same time 2 versions of Coq on the same sources.

Example: b.v depending on a.v

  coq8.6/bin/coqc -R out8.6 Test src/a.v -o out8.6/a.vo
  coq8.6/bin/coqc -R out8.6 Test src/b.v -o out8.6/b.vo

  coq8.7/bin/coqc -R out8.7 Test src/a.v -o out8.7/a.vo
  coq8.7/bin/coqc -R out8.7 Test src/b.v -o out8.7/b.vo