Proofview: extensions for backtracking eauto
unshelve_goals is used to correctly register dependent
subgoals that have to be solved. Resolution may fail to
do so using hints, so we have to put them back as goals
in that case. The shelf is a good interface for doing that.

unifiable can be used outside proofview to detect dependencies
between goals. This might be better in another module.