About printing of traces of failures while calling ltac code.
An Ltac trace printing mechanism was introduced in 8.4 which was
inadvertedly modified by a series of commits such as 8e10368c3,
91f44f1da7a, ...

It was also sometimes buggy, iirc, when entering ML tactics which
themselves were calling ltac code.

It got really bad in 8.5 as in:

Tactic Notation "f" constr(x) := apply x. Ltac g x := f x.
Goal False.
idtac; f I. (* bad location reporting *)
g I.        (* was referring to tactic name "Top.Top#<>#1" *)

which this commit fixes.

I don't have a clear idea of what would be the best ltac tracing
mechanism, but to avoid it to be broken without being noticed, I
started to add some tests.

Eventually, it might be worth that an Ltac expert brainstrom on it!