Preventive compatibility fixes for flushing.
In pre 8.6, `Pp` provided its own reimplementation of
`Pervasives.flush_all`, with different semantics.

Unfortunately, with the removal of `Pp.flush_all` in #179, a couple of
points were silently switched to the `Pervasives` version, which may
lead to some subtle printing differences.

As a preventive measure, we restore the same semantics for these parts
of the codebase.

Note that we don't re-introduce Coq's `flush_all` for several reasons:

- Consumers of the logging API should not mess with flushing and
  Formatters as this is backend dependent (i.e: when in IDEs).

  Use of `Format` should be fully encapsulated if we want some hope of
  IDEs taking full control.

- As used, the old semantics of `flush_all` were fragile.