[stm] Extend the `Stm.add` interface.
We give the `Stm.add` user more control over the document and in
particular about location reporting. The main changes on `Stm.add` are:

- We propagate the location information generated by the parser that
  signals the end of the parsed sentence. This way, consumers have an
  accurate way of detecting sentence information.

- We remove the indentation guessing from the `Stm`, as well as the
  indentation and length fields; this information is contained now in
  the location.

- `Stm.add` now takes a `parsable`. This allows for more accurate
  position reporting as the caller has control over the lexer now.

  This improves incremental document building with successive `Stm.add`
  calls.

We don't solve yet the duplication of parsing entry points among `Stm`
and `Vernac`.

We also add `stm` entries to the `.merlin` file, they were missing and
causing merlin to fail.