Prevent unsafe overwriting of Required modules by toplevel library.
In coqtop, one could do for instance:

Require Import Top. (* Where Top contains a Definition b := true *)
Lemma bE : b = true. Proof. reflexivity. Qed.

Definition b := false.

Lemma bad : False. Proof. generalize bE; compute; discriminate. Qed.

That proof could however not be saved because of the circular dependency check.

As a stopgap, we register the toplevel library in the Nametab. This way, this
proof of false is prevented in the same way as is a similar one with a library
whose logical name is A.B and another library whose logical name is A and which
contains a module B.

A better long term fix for these issues would not rely on Nametab, which is
effectively in the TCB today.